POAP-88 WIP Refactor folders
