+++ server/src/eth/VotePoap.d.ts
